Search: Remove zoekt.Options struct
