+++ internal/search/BUILD.bazel
@@ -26,10 +26,13 @@ go_library(
         "//internal/search/query",
         "//internal/search/result",
         "//internal/trace",
+        "//internal/trace/ot",
+        "//internal/trace/policy",
         "//internal/types",
         "//lib/errors",
         "//schema",
         "@com_github_grafana_regexp//:regexp",
+        "@com_github_opentracing_opentracing_go//:opentracing-go",
         "@com_github_sourcegraph_log//:log",
         "@com_github_sourcegraph_zoekt//:zoekt",
         "@com_github_sourcegraph_zoekt//query",
@@ -43,13 +46,20 @@ go_test(
     srcs = [
         "alert_test.go",
         "repo_status_test.go",
+        "types_test.go",
     ],
     embed = [":search"],
     deps = [
         "//internal/api",
+        "//internal/conf",
+        "//internal/search/filter",
+        "//internal/search/limits",
         "//internal/search/query",
+        "//schema",
         "@com_github_google_go_cmp//cmp",
         "@com_github_google_go_cmp//cmp/cmpopts",
+        "@com_github_sourcegraph_log//logtest",
+        "@com_github_sourcegraph_zoekt//:zoekt",
         "@com_github_stretchr_testify//require",
     ],
 )
