+++ internal/search/zoekt/BUILD.bazel
@@ -28,13 +28,10 @@ go_library(
         "//internal/search/result",
         "//internal/search/streaming",
         "//internal/trace",
-        "//internal/trace/ot",
-        "//internal/trace/policy",
         "//internal/types",
         "//lib/errors",
         "@com_github_go_enry_go_enry_v2//:go-enry",
         "@com_github_grafana_regexp//:regexp",
-        "@com_github_opentracing_opentracing_go//:opentracing-go",
         "@com_github_roaringbitmap_roaring//:roaring",
         "@com_github_sourcegraph_log//:log",
         "@com_github_sourcegraph_zoekt//:zoekt",
@@ -54,19 +51,16 @@ go_test(
     embed = [":zoekt"],
     deps = [
         "//internal/api",
-        "//internal/conf",
         "//internal/search",
         "//internal/search/backend",
         "//internal/search/filter",
         "//internal/search/job",
-        "//internal/search/limits",
         "//internal/search/query",
         "//internal/search/result",
         "//internal/search/streaming",
         "//internal/trace",
         "//internal/types",
         "//lib/errors",
-        "//schema",
         "@com_github_google_go_cmp//cmp",
         "@com_github_google_go_cmp//cmp/cmpopts",
         "@com_github_grafana_regexp//:regexp",
